build: default config for max lines eslint rule
