+++ src/service/s3/bucket/website/s3-bucket-website.loc.test.ts
@@ -117,43 +117,6 @@ describe("Serve simulated S3 Bucket static website on localhost", () => {
     assertIdentical(await res.text(), "<h1>Docs index</h1>");
   });
 
-  it("redirects to a slash-terminated folder URL when a folder index document exists", async () => {
-    const simS3 = simAws.region("eu-west-2").s3();
-
-    await simS3.createBucket(
-      new CreateBucketCommand({ Bucket: "folder-index-redirect-site" }),
-    );
-    await simS3.putObject(
-      new PutObjectCommand({
-        Bucket: "folder-index-redirect-site",
-        Key: "docs/index.html",
-        Body: "<h1>Docs index</h1>",
-        ContentType: "text/html; charset=utf-8",
-      }),
-    );
-    await simS3.putBucketWebsite(
-      new PutBucketWebsiteCommand({
-        Bucket: "folder-index-redirect-site",
-        WebsiteConfiguration: {
-          IndexDocument: {
-            Suffix: "index.html",
-          },
-        },
-      }),
-    );
-
-    const res = await fetch(
-      `http://folder-index-redirect-site.s3-website.eu-west-2.sim-aws.localhost:${srv.port}/docs`,
-      { redirect: "manual" },
-    );
-
-    assertIdentical(res.status, 301);
-    assertIdentical(
-      res.headers.get("location"),
-      `http://folder-index-redirect-site.s3-website.eu-west-2.sim-aws.localhost:${srv.port}/docs/`,
-    );
-  });
-
   it("serves the configured error document body with HTTP 404", async () => {
     const simS3 = simAws.region("eu-west-2").s3();
 
@@ -192,120 +155,5 @@ describe("Serve simulated S3 Bucket static website on localhost", () => {
       "text/html; charset=utf-8",
     );
     assertIdentical(await res.text(), "<h1>Not found</h1>");
-  });
-
-  it("redirects missing objects with a matching HTTP error routing rule", async () => {
-    const simS3 = simAws.region("eu-west-2").s3();
-
-    await simS3.createBucket(
-      new CreateBucketCommand({ Bucket: "missing-redirect-site" }),
-    );
-    await simS3.putBucketWebsite(
-      new PutBucketWebsiteCommand({
-        Bucket: "missing-redirect-site",
-        WebsiteConfiguration: {
-          IndexDocument: {
-            Suffix: "index.html",
-          },
-          RoutingRules: [
-            {
-              Condition: {
-                HttpErrorCodeReturnedEquals: "404",
-              },
-              Redirect: {
-                HttpRedirectCode: "302",
-                ReplaceKeyWith: "not-found.html",
-              },
-            },
-          ],
-        },
-      }),
-    );
-
-    const res = await fetch(
-      `http://missing-redirect-site.s3-website.eu-west-2.sim-aws.localhost:${srv.port}/missing.html`,
-      { redirect: "manual" },
-    );
-
-    assertIdentical(res.status, 302);
-    assertIdentical(
-      res.headers.get("location"),
-      `http://missing-redirect-site.s3-website.eu-west-2.sim-aws.localhost:${srv.port}/not-found.html`,
-    );
-  });
-
-  it("resolves an index document before checking HTTP error routing rules", async () => {
-    const simS3 = simAws.region("eu-west-2").s3();
-
-    await simS3.createBucket(
-      new CreateBucketCommand({ Bucket: "index-before-rule-site" }),
-    );
-    await simS3.putObject(
-      new PutObjectCommand({
-        Bucket: "index-before-rule-site",
-        Key: "docs/index.html",
-        Body: "<h1>Docs index</h1>",
-        ContentType: "text/html; charset=utf-8",
-      }),
-    );
-    await simS3.putBucketWebsite(
-      new PutBucketWebsiteCommand({
-        Bucket: "index-before-rule-site",
-        WebsiteConfiguration: {
-          IndexDocument: {
-            Suffix: "index.html",
-          },
-          RoutingRules: [
-            {
-              Condition: {
-                HttpErrorCodeReturnedEquals: "404",
-              },
-              Redirect: {
-                ReplaceKeyWith: "not-found.html",
-              },
-            },
-          ],
-        },
-      }),
-    );
-
-    const res = await fetch(
-      `http://index-before-rule-site.s3-website.eu-west-2.sim-aws.localhost:${srv.port}/docs/`,
-      { redirect: "manual" },
-    );
-
-    assertIdentical(res.status, 200);
-    assertIdentical(res.headers.get("location"), null);
-    assertIdentical(await res.text(), "<h1>Docs index</h1>");
-  });
-
-  it("redirects all requests when RedirectAllRequestsTo is configured", async () => {
-    const simS3 = simAws.region("eu-west-2").s3();
-
-    await simS3.createBucket(
-      new CreateBucketCommand({ Bucket: "redirect-all-site" }),
-    );
-    await simS3.putBucketWebsite(
-      new PutBucketWebsiteCommand({
-        Bucket: "redirect-all-site",
-        WebsiteConfiguration: {
-          RedirectAllRequestsTo: {
-            HostName: "example.test",
-            Protocol: "https",
-          },
-        },
-      }),
-    );
-
-    const res = await fetch(
-      `http://redirect-all-site.s3-website.eu-west-2.sim-aws.localhost:${srv.port}/docs/page.html`,
-      { redirect: "manual" },
-    );
-
-    assertIdentical(res.status, 301);
-    assertIdentical(
-      res.headers.get("location"),
-      "https://example.test/docs/page.html",
-    );
   });
 });
