+++ src/service/s3/storage/s3-filesystem-storage.ts
@@ -1,8 +1,8 @@
 import type { SimS3BucketStorage } from "./s3-bucket-storage.js";
 import { SimS3Object, SimS3ObjectMetadata } from "../object/s3-object.js";
 import path from "node:path";
-import { homedir } from "node:os";
 import { mkdir, readdir, readFile, stat, writeFile } from "node:fs/promises";
+import { FilesystemS3StorageSafety } from "./s3-filesystem-safety.js";
 
 interface FilesystemS3BucketStorageProps {
   readonly directoryPath: string;
@@ -19,12 +19,13 @@ interface FilesystemS3BucketStorageProps {
  */
 export class FilesystemS3BucketStorage implements SimS3BucketStorage {
   private readonly directoryPath: string;
-  private readonly allowedDirectoryNames: readonly string[];
+  private readonly safety: FilesystemS3StorageSafety;
 
   constructor(props: FilesystemS3BucketStorageProps) {
-    const { allowedDirectoryNames = defaultAllowedDirectoryNames } = props;
-    this.allowedDirectoryNames = allowedDirectoryNames;
-    this.assertSafeDirectoryPath(props.directoryPath);
+    this.safety = new FilesystemS3StorageSafety({
+      allowedDirectoryNames: props.allowedDirectoryNames,
+    });
+    this.safety.assertSafeDirectoryPath(props.directoryPath);
     this.directoryPath = path.resolve(props.directoryPath);
   }
 
@@ -32,8 +33,8 @@ export class FilesystemS3BucketStorage implements SimS3BucketStorage {
    * Get a simulated Object from a file in the directory.
    */
   async getObject(key: string): Promise<SimS3Object | undefined> {
-    if (!this.isAllowedObjectKeyExtension(key)) {
-      this.assertSafeObjectKeyPath(key);
+    if (!this.safety.isAllowedObjectKeyExtension(key)) {
+      this.safety.assertSafeObjectKeyPath(key);
       return undefined;
     }
 
@@ -105,7 +106,7 @@ export class FilesystemS3BucketStorage implements SimS3BucketStorage {
   }
 
   private filePathForObjectKey(key: string): string {
-    this.assertSafeObjectKey(key);
+    this.safety.assertSafeObjectKey(key);
 
     const filePath = path.resolve(this.directoryPath, key);
 
@@ -164,7 +165,7 @@ export class FilesystemS3BucketStorage implements SimS3BucketStorage {
           .split(path.sep)
           .join("/");
 
-        if (this.isAllowedObjectKeyExtension(key)) {
+        if (this.safety.isAllowedObjectKeyExtension(key)) {
           keys.push(key);
         }
       }
@@ -265,121 +266,4 @@ export class FilesystemS3BucketStorage implements SimS3BucketStorage {
       }
     }
   }
-
-  private assertSafeDirectoryPath(directoryPath: string): void {
-    if (!path.isAbsolute(directoryPath)) {
-      throw new Error(
-        `Filesystem S3 storage directory path must be absolute: ${directoryPath}`,
-      );
-    }
-
-    if (this.pathContainsParentDirectorySegment(directoryPath)) {
-      throw new Error(
-        `Filesystem S3 storage directory path must not contain '..': ${directoryPath}`,
-      );
-    }
-
-    const resolvedDirectoryPath = path.resolve(directoryPath);
-    const parsedDirectoryPath = path.parse(resolvedDirectoryPath);
-
-    if (resolvedDirectoryPath === parsedDirectoryPath.root) {
-      throw new Error(
-        `Filesystem S3 storage directory path must not be a filesystem root: ${directoryPath}`,
-      );
-    }
-
-    if (resolvedDirectoryPath === path.resolve(homedir())) {
-      throw new Error(
-        `Filesystem S3 storage directory path must not be the user home directory: ${directoryPath}`,
-      );
-    }
-
-    const directoryName = path.basename(resolvedDirectoryPath);
-
-    if (!this.allowedDirectoryNames.includes(directoryName)) {
-      throw new Error(
-        `Filesystem S3 storage directory name must be one of: ${this.allowedDirectoryNames.join(
-          ", ",
-        )}. Got: ${directoryName}`,
-      );
-    }
-  }
-
-  private assertSafeObjectKey(key: string): void {
-    this.assertSafeObjectKeyPath(key);
-
-    if (!this.isAllowedObjectKeyExtension(key)) {
-      throw new Error(`S3 Object key has unsupported file extension: ${key}`);
-    }
-  }
-
-  private assertSafeObjectKeyPath(key: string): void {
-    if (path.isAbsolute(key)) {
-      throw new Error(`S3 Object key must not be an absolute path: ${key}`);
-    }
-
-    if (this.pathContainsParentDirectorySegment(key)) {
-      throw new Error(`S3 Object key must not contain '..': ${key}`);
-    }
-  }
-
-  private pathContainsParentDirectorySegment(value: string): boolean {
-    return value.split(/[\\/]/u).includes("..");
-  }
-
-  private isAllowedObjectKeyExtension(key: string): boolean {
-    return allowedObjectFileExtensions.has(path.extname(key).toLowerCase());
-  }
 }
-
-/**
- * Options for filesystem-based simulated S3 Bucket storage.
- */
-export interface FilesystemS3BucketStorageOptions {
-  /**
-   * Directory names that are safe to use as filesystem storage roots.
-   */
-  readonly allowedDirectoryNames?: readonly string[];
-}
-
-/**
- * Default directory names that are safe to use as filesystem storage roots.
- */
-const defaultAllowedDirectoryNames = [
-  "assets",
-  "build",
-  "dist",
-  "out",
-  "public",
-  "static",
-  "www",
-] as const;
-
-/**
- * Cautious list of allowed file extensions for simulated S3 objects. This is to
- * try and avoid reading or writing other files that might be unsafe.
- */
-const allowedObjectFileExtensions = new Set([
-  ".css",
-  ".eot",
-  ".gif",
-  ".htm",
-  ".html",
-  ".ico",
-  ".jpeg",
-  ".jpg",
-  ".js",
-  ".json",
-  ".map",
-  ".mjs",
-  ".otf",
-  ".png",
-  ".svg",
-  ".ttc",
-  ".ttf",
-  ".txt",
-  ".webp",
-  ".woff",
-  ".woff2",
-  ".xml",
-]);
